Stock Search Website/backend/requirements.txt
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonmerge==1.9.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
python-dateutil==2.8.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
urllib3==2.2.0
Werkzeug==3.0.1
